+++ Changelog.md
@@ -1,6 +1,6 @@
 Flan 1.7.7
 ================
-- Prevent moss bonemealing if too close to a claim without permission
+- Prevent moss/grassblock/nylium bonemealing if too close to a claim without permission
 
 Flan 1.7.6
 ================